refactoring: DTO, Entity 구분
1. CarNumber 이동

Resolves:
Ref:
Related to:
