+++ demo/src/main/java/com/example/demo/repo/CarNumberRepo.java
@@ -1,6 +1,6 @@
 package com.example.demo.repo;
 
-import com.example.demo.DTO.CarNumber;
+import com.example.demo.Entity.CarNumber;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
